documentation now generates correctly on local machine.
