+++ Examples/other_examples.jl
@@ -229,13 +229,11 @@ function prism_refraction()
     end
     raygen = Emitters.Sources.RayListSource(rays)
     # draw the result
-    Vis.drawtracerays(sys, raygenerator=raygen, test=true, trackallrays=true)
 end
 
 function fresnel(convex=true; kwargs...)
     lens = FresnelLens(Examples_N_BK7, 0.0, convex ? 15.0 : -15.0, 1.0, 8.0, 0.8, conic=0.1)
     sys = CSGOpticalSystem(LensAssembly(lens()), Rectangle(15.0, 15.0, SVector(0.0, 0.0, 1.0), SVector(0.0, 0.0, -25.0), interface=opaqueinterface()))
-    Vis.drawtracerays(sys; test=true, trackallrays=true, numdivisions=30, kwargs...)
 end
 
 """This example no longer works correctly. The visualization code needs to be updated to support RayListSource"""
@@ -314,22 +312,5 @@ function eyetrackHOE(nrays=5000, det=false, showhead=true, zeroorder=false; kwar
     source = Emitters.Sources.RayListSource(rays)
 
     sys = CSGOpticalSystem(LensAssembly(obj, cornea, camlens, cambarrel, camap), camdet, 800, 800)
-    if det
-        Vis.show(OpticSim.traceMT(sys, source))
-    else
-        Vis.drawtracerays(sys; raygenerator=source, trackallrays=true, kwargs...)
-        # for θ in 0:(π / 6):(2π)
-        #     ledloc = SVector(20 * cos(θ) + offset[1], 0 + offset[2], 15 * sin(θ) + offset[3])
-        #     Vis.draw!(leaf(Sphere(1.0), translation(ledloc...)), color = :red)
-        # end
-        for d in dirs
-            # Vis.draw!(leaf(Sphere(1.0), translation((corneavertex - 10 * d)...)), color = :red)
-            Vis.draw!((corneavertex - 50 * d, corneavertex), color=:red)
-        end
-        if showhead
-            Vis.draw!(joinpath(@__DIR__, "../../OBJ/glasses.obj"), scale=100.0, transform=Transform(OpticSim.rotmatd(90, 0, 0), [27.0, 45.0, -8.0]), color=:black)
-            Vis.draw!(joinpath(@__DIR__, "../../OBJ/femalehead.obj"), scale=13.0, transform=Transform(OpticSim.rotmatd(0, 0, 180), [27.0, 105.0, -148.0]), color=:white)
-        end
-        Vis.display()
-    end
+
 end
