+++ docs/Project.toml
@@ -1,4 +1,5 @@
 [deps]
+AGFFileReader = "36dd7d7d-c279-4b5c-a024-2ba417cea492"
 CodeTracking = "da1fd8a2-8d9e-5ec2-8556-3022fb5608a2"
 DataFrames = "a93c6f00-e57d-5684-b7b6-d8193f3e46c0"
 Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
