+++ docs/src/emitters.md
@@ -11,7 +11,7 @@ The **basic emitter** ([`Source`](@ref sources)) is constructed as a combination
 - [`Rays Origins Distribution`](@ref rays_origins_distribution)
 - [`Rays Directions Distribution`](@ref rays_directions_distribution)
 
-The [`OpticSim`](index.html) package comes with various implementations of each of these basic elements:
+The `OpticSim` package comes with various implementations of each of these basic elements:
 - Spectrum - the **generate** interface returns a tuple (power, wavelength)
   * [`Emitters.Spectrum.Uniform`](@ref) - A flat spectrum bounded (default: from 450nm to 680nm). the range is sampled uniformly.
   * [`Emitters.Spectrum.DeltaFunction`](@ref) - Constant wave length.
@@ -36,9 +36,7 @@ The [`OpticSim`](index.html) package comes with various implementations of each 
 **Note**: All of the examples on this page assume that the following statement was executed:
 
 ```@example 
-dummy = "switching to WGLMakie in order to produce interactive figures with Makie"  # hide
 using OpticSim, OpticSim.Geometry, OpticSim.Emitters                                # hide
-OpticSim.NotebooksUtils.SetDocsBackend("Web")                                       # hide
 ```
 
 ```@example
@@ -48,44 +46,39 @@ using OpticSim, OpticSim.Geometry, OpticSim.Emitters
 ### Simple functions for creating commonly used emitters
 Many optical systems by convention have their optical axis parallel to the z axis. These utility functions provide a simple interface to the Emitters package to create emitters that direct their rays in the negative z direction, toward the entrance of the optical system.
 
-```@docs
+```@docs; canonical = false
 Emitters.pointemitter
 ```
 
 ```@example
 using OpticSim, OpticSim.Geometry, OpticSim.Emitters # hide
 pt = Emitters.pointemitter([0.0,0.0,.5],.3)
-Vis.draw(pt, debug=true, resolution = (800, 600))
 ```
 
-```@docs
+```@docs; canonical = false
 Emitters.collimatedemitter
 ```
 
 ```@example
 using OpticSim, OpticSim.Geometry, OpticSim.Emitters # hide
 pt = Emitters.collimatedemitter([0.0,0.0,1.0],.5)
-Vis.draw(pt, debug=true, resolution = (800, 600))
 ```
 
 ### Point origin with various Direction distributions
 ```@example
 using OpticSim, OpticSim.Geometry, OpticSim.Emitters # hide
 src = Sources.Source(origins=Origins.Point(), directions=Directions.RectGrid(π/4, π/4, 15, 15))
-Vis.draw(src, debug=true, resolution = (800, 600))
 ```
 
 ```@example
 using OpticSim, OpticSim.Geometry, OpticSim.Emitters # hide
 src = Sources.Source(origins=Origins.Point(), directions=Directions.UniformCone(π/6, 1000))
-Vis.draw(src, debug=true, resolution = (800, 600))
 ```
 
 
 ```@example
 using OpticSim, OpticSim.Geometry, OpticSim.Emitters # hide
 src = Sources.Source(origins=Origins.Point(), directions=Directions.HexapolarCone(π/6, 10))
-Vis.draw(src, debug=true, resolution = (800, 600))
 ```
 
 
@@ -94,14 +87,12 @@ Vis.draw(src, debug=true, resolution = (800, 600))
 ```@example
 using OpticSim, OpticSim.Geometry, OpticSim.Emitters # hide
 src = Sources.Source(origins=Origins.RectGrid(1.0, 1.0, 10, 10), directions=Directions.Constant())
-Vis.draw(src, debug=true, resolution = (800, 600))
 ```
 
 
 ```@example
 using OpticSim, OpticSim.Geometry, OpticSim.Emitters # hide
 src = Sources.Source(origins=Origins.Hexapolar(5, 1.0, 2.0), directions=Directions.Constant())
-Vis.draw(src, debug=true, resolution = (800, 600))
 ```
 
 
@@ -116,7 +107,6 @@ src = Sources.Source(
 	directions=Directions.RectGrid(π/6, π/6, 15, 15),   # RectGrid Directions
 	power=AngularPower.Cosine(10.0)                     # Cosine Angular Power 
 )
-Vis.draw(src, debug=true, resolution = (800, 600))
 ```
 
 ```@example
@@ -126,7 +116,6 @@ src = Sources.Source(
 	directions=Directions.HexapolarCone(π/6, 10),       # HexapolarCone Directions
 	power=AngularPower.Gaussian(2.0, 2.0)               # Gaussian Angular Power 
 )
-Vis.draw(src, debug=true, resolution = (800, 600))
 ```
 
 ### Composite Sources - Display Example
@@ -163,22 +152,19 @@ end
 Tr = Transform(Vec3(0.0, 0.0, 0.0))
 my_display = Sources.CompositeSource(Tr, pixels)
 
-Vis.draw(my_display)                                                # render the display - nothing but the origins primitives
-rays = AbstractArray{OpticalRay{Float64, 3}}(collect(my_display))   # collect the rays generated by the display
-Vis.draw!(rays)                                                     # render the rays 
+rays = AbstractArray{OpticalRay{Float64, 3}}(collect(my_display))   # collect the rays generated by the display                                                   # render the rays 
 ```
 
 
 ```@example 
 dummy = "switching Back to the GLMakie to allow the rest of the pages to work with static figures"  # hide
 using OpticSim, OpticSim.Geometry, OpticSim.Emitters                                                # hide
-OpticSim.NotebooksUtils.SetDocsBackend("Static")                                                    # hide
 ```
 
 
 ## [Spectrum](@id spectrum)
 
-```@docs
+```@docs; canonical = false
 Emitters.Spectrum.Uniform
 Emitters.Spectrum.DeltaFunction
 Emitters.Spectrum.Measured
@@ -187,7 +173,7 @@ Emitters.Spectrum.spectrumpower
 
 ## [Angular Power Distribution](@id angular_power_distribution)
 
-```@docs
+```@docs; canonical = false
 Emitters.AngularPower.Lambertian
 Emitters.AngularPower.Cosine
 Emitters.AngularPower.Gaussian
@@ -195,7 +181,7 @@ Emitters.AngularPower.Gaussian
 
 ## [Rays Origins Distribution](@id rays_origins_distribution)
 
-```@docs
+```@docs; canonical = false
 Emitters.Origins.Point
 Emitters.Origins.RectUniform
 Emitters.Origins.RectGrid
@@ -204,7 +190,7 @@ Emitters.Origins.Hexapolar
 
 ## [Rays Directions Distribution](@id rays_directions_distribution)
 
-```@docs
+```@docs; canonical = false
 Emitters.Directions.Constant
 Emitters.Directions.RectGrid
 Emitters.Directions.UniformCone
@@ -213,7 +199,7 @@ Emitters.Directions.HexapolarCone
 
 ## [Sources](@id sources)
 
-```@docs
+```@docs; canonical = false
 Emitters.Sources.Source
 Emitters.Sources.CompositeSource
 ```
