+++ docs/src/lenses.md
@@ -5,7 +5,7 @@
 A number of helper functions are provided to make constructing simple lenses easier.
 Firstly ordinary thick lenses:
 
-```@docs
+```@docs; canonical = false
 SphericalLens
 ConicLens
 AsphericLens
@@ -14,7 +14,7 @@ FresnelLens
 
 As well as idealized lenses:
 
-```@docs
+```@docs; canonical = false
 ParaxialLens
 ```
 
@@ -24,7 +24,7 @@ We also have some holographic elements implemented, note that these have not bee
 
 It is relatively simple to extend the existing code to add these kinds of specialized surfaces providing a paired [`OpticalInterface`](@ref) subclass is also defined. In this case the `WrapperSurface` can often serve as a suitable base for extension.
 
-```@docs
+```@docs; canonical = false
 WrapperSurface
 ThinGratingSurface
 HologramSurface
@@ -35,7 +35,7 @@ MultiHologramSurface
 
 Eye models are often very useful in simulation of head mounted display systems. We have two models implemented currently.
 
-```@docs
-ModelEye
-ArizonaEye
+```@docs; canonical = false
+OpticSim.Data.ModelEye
+OpticSim.Data.ArizonaEye
 ```
