--- docs/src/notebooksutils.md
@@ -1,13 +1,0 @@
-# Notebook utilities
-
-```@meta
-CurrentModule = NotebooksUtils
-```
-
-[TODO]
-
-```@docs
-run_sample
-SetBackend
-run
-```
