+++ docs/src/systems.md
@@ -4,7 +4,7 @@
 
 All systems are made up of a `LensAssembly` which contains all the optical components in the system, excluding any sources (see [Emitters](@ref)) and the detector (see below).
 
-```@docs
+```@docs; canonical = false
 LensAssembly
 ```
 
@@ -12,7 +12,7 @@ LensAssembly
 
 The detector image is stored within the system as a `HierarchicalImage` for memory efficiency.
 
-```@docs
+```@docs; canonical = false
 HierarchicalImage
 OpticSim.reset!
 OpticSim.sum!
@@ -22,7 +22,7 @@ OpticSim.sum!
 
 There are two types of `AbstractOpticalSystem` which can be used depending on the requirements.
 
-```@docs
+```@docs; canonical = false
 AbstractOpticalSystem
 CSGOpticalSystem
 AxisymmetricOpticalSystem
@@ -38,7 +38,7 @@ semidiameter
 
 We can trace an individual [`OpticalRay`](@ref) through the system (or directly through a [`LensAssembly`](@ref)), or we can trace using an [`OpticalRayGenerator`](@ref) to create a large number of rays.
 
-```@docs
+```@docs; canonical = false
 trace
 traceMT
 tracehits
