+++ src/OpticSim.jl
@@ -25,10 +25,10 @@ include("utilities.jl")
 
 
 
-include("Data/Data.jl")
+
 include("Geometry/Geometry.jl")
 include("Optical/Optical.jl")
-
+include("Data/Data.jl")
 
 #initialize these caches here so they will get the correct number of threads from the load time environment, rather than the precompile environment. The latter happens if the initialization happens in the const definition. If the precompile and load environments have different numbers of threads this will cause an error.
 function __init__()
